Lint removal in actions for zitilib models.
