--- zitilib/characterization/actions/logs.go
@@ -1,33 +1,0 @@
-/*
-	Copyright 2020 NetFoundry, Inc.
-
-	Licensed under the Apache License, Version 2.0 (the "License");
-	you may not use this file except in compliance with the License.
-	You may obtain a copy of the License at
-
-	https://www.apache.org/licenses/LICENSE-2.0
-
-	Unless required by applicable law or agreed to in writing, software
-	distributed under the License is distributed on an "AS IS" BASIS,
-	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-	See the License for the specific language governing permissions and
-	limitations under the License.
-*/
-
-package zitilib_characterization_actions
-
-import (
-	"github.com/netfoundry/fablab/kernel/model"
-	"github.com/netfoundry/fablab/zitilib/actions/logs"
-)
-
-func NewLogsAction() model.ActionBinder {
-	action := &logsAction{}
-	return action.bind
-}
-
-func (self *logsAction) bind(_ *model.Model) model.Action {
-	return logs.Logs()
-}
-
-type logsAction struct{}
