+++ zitilib/console/client.go
