+++ zitilib/console/mgmt.go
